requirements.txt
pymongo
shapely.geometry
pyproj